requirements.txt
# Web + API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Dashboard
dash>=2.17.0
plotly>=5.22.0
dash-bootstrap-components>=1.6.0
dash-extensions>=1.0.14
Flask-Caching>=2.3.0

# Desktop
pywebview>=5.0.5
pyinstaller>=6.6.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.7
httpx>=0.27.0

# Utilities
pandas>=2.2.2
python-dateutil>=2.9.0
